Fix broken unit test
I had to add a line to the constructor that would set the current time
of the fake packet, but forgot to check before doing so that the
pointer was non-null.

This broke one unit test, and of course because I was hacking on stuff
that couldn't be unit tested, I forgot to run the unit tests before
committing.