Add a comment
Setup_object_data always did an odd thing where it discards any
user-provided lat/lon if one was doing a Modify operation from a
dialog on an existing object, if that object had a course and speed
--- instead it would use the dead-reckoned position.

I preserved this behavior and tested that it worked as always, but I
am not sure I like that we do that.  A user modifying the position of
an object using the dialog would not expect that modification to be
ignored just because the previous report had a speed.

So I have made sure to comment on this in the header to
Setup_object_data.