Begin refactor of Setup_item_data
This is a straight cut/paste job out of Setup_object_data, with only
the modifications necessary to replace the checking of item name and
passing a "0" to construct_object_item_data_row, signfying item
instead of object.

Further refactor is needed --- Create_object_item_tx_string already
checks the validity of the object/item name, so it is not necessary to
be redoing it in Setup_object_data and Setup_item_data.   That's for a
later commit.

As committed here, the code works and creates correctly formatted item
packets for transmit.