Further simplify object/item deletion code
Add an "is_killed" argument to Setup_object_item_data.  This argument
is simply passed down to construct_object_item_data_row which will set
or unset the ST_ACTIVE bit in the DataRow flags field.  That flag is
in turn used by create_object_item_tx_string to create a packet string
for a killed object or item.

Doing this, make the two functions that call Setup_object_item_data
for killed packets pass a 1 for is_killed instead of hacking on the
packet themselves to change the appropriate character of the
non-killed packet to a "_".

Closes #265