Remove the "travis" configuration file
Some years ago, Github paired with another system called Travis to
provide CI builds.  It was free for a while at least for open source
projects, and we jumped on that action to do automated build testing
of commits.

Free Travis support ended some time ago and we went a long time with
no CI.

CI is now provided by Github Actions, and we're using that to test on
linux, mac, and freebsd.

I'm deleting the old travis configuration file, `.travis.yml`.