Move a line of code inside a conditional
One of the last things I did to construct_object_item_data_row was
make it set the "sec_heard" field to the current time.  The first time
I did it I did it right before the return, without checking that the
datarow pointer was valid.  That caused a unit test failure so I
quickly slammed in a test of valid pointer right before doing the
setting.

However, the *entire* block of code just above it is already
conditional on the pointer being valid, so the right thing to have
done would have been to put that line inside the block instead of
outside.

This commit just moves the setting of sec_heard inside the block of
code where it belonged in the first place.