Add a unit tests for object/item formatters with overlay symbol
One of the main reasons I got off on this whole refactor business for
objects/items was as a consequence of #239, where it was reported that
Xastir was not properly formatting compressed posits if the station
had a symbol with numeric overlay.

My first test confirming that bug was to try creating an object with
a numeric overlay and it didn't work, either.

This commit actually adds unit tests for exactly that issue.

There are now two tests of object/item formatters in objects_utils.c when
asked to format a packet with a numeric overlay, in both compressed and
uncompressed form.

The "5" for the overlay should be replaced by an 'f' in the compressed
packet, and these new test verifies that that happens.

There are also two new tests of construct_object_item_data_row and
Create_object_item_tx_string (in objects.c) for the same two cases: an
object with a numeric overlay, compressed and uncompressed.

The object_util.c tests check the low-level string formatters.

The objects.c tests check the high-level code that actually does the
creation of objects from dialog box values (that is, when *passed*
dialog box values --- the functions don't read the dialogs
themselves).  Since the high-level functions call the low level
functions, and it's actually the low-level functions that handle the
formatting, this is sort of overkill, but overkill is good when we're
talking about unit tests.

All that refactor inspired by one simple issue, and the unit tests
didn't even verify that the issue was addressed (largely because the
issue was fixed long before the refactor started, and memory is
short).  Better late than never.