Fix broken logic if dbfawk sigs not found
There was a possibility of broken map rendering if no dbfawk
signatures are found (at all) in the config directory.  A conditional
that tested for whether any were found had no "else" clause, but this
could lead to no default signature being used and map rendering to be
broken.

Now, if NO dbfawk signatures are found in the config directory, we at
least always use the default dbfawk signature.

This leads to UGLY map rendering, but at least it doesn't fall through
and start using uninitialized variables like it could have before.