Remove massive, obsolete comment
draw_shapefile_map had a massive comment block at the top that
described how it decoded alerts and mapped them onto shapefile fields
for an extremely ancient version of NWS shapefiles, when they had to
be hand-coded.

Since all alert handling is now done through dbfawk files that provide
a "key" variable for matching the alert zone, and NWS has changed the
shapefile fields many, many times over the years, all of that stuff is
obsolete.  We have to keep updating dbfawk files every time NWS shifts
the ground we're built on, and we do that regularly.

There were also comments that said we had to hand-code for every new
shapefile source, which hasn't been true for almost 20 years (as long
as dbfawk is compiled in, which it now MUST be for shapefile support).

I'm removing those comments and adding a few lines pointing at other
parts of the code that handle weather alerts.