Remove relics of old hard-coded quad overlay rendering
The hand-coded shapefile rendering in draw_shapefile_map had special
cases for every specific type of shapefile that was supported,
including a large "quad overlay" that displayed the outlines of every
USGS 1:24000 quad (the link pointing to where to find this shapefile is
long dead).

But since 2004, there has been a DBFAWK file that lets this same
shapefile render without the hard-coding, and the hard-coding was
mostly removed in commit 3a9b6e1 in 2019.

But I removed the hardcoded stuff using the Unix utility "unifdef"
that simply yanked out all the stuff that had been ifdeffed out if
dbfawk wasn't compiled in.   There were several chunks of hard-coded
stuff to handle the 24k shapefile that were just dead code now because
the variable on which they were all conditioned was never set to
anything other than zero.

This commit removes all of those chunks.

It also removes two blocks of code that had been ifdeffed out with
"ifdef notdef" as part of dbfawk debugging in 2003, and never cleaned
out.

closes #271