Remove two bits of non-dbfawk cruft
In commit 3a9b6e1 I ostensibly removed the non-dbfawk shapefile
rendering, but when I did it I did so only by using the Unix utility
"unifdef" that removed all the stuff inside "ifndef DBFAWK".

There were several lines that weren't inside that ifdef, but that had
no real effect.  Those were removed in the previous commit.

I found two more.  One is a comment that referred to things inside an
ifdef, but was left outside the ifdef.  The other was obviously an
accidental duplication of a single line of code both before and after
the ifdef, setting the foreground color. When the ifdef was removed,
these two identical lines were separated only by that comment.

I have removed the comment and the duplicate line.

Both of those things should have been removed when I was working on