Make object/item datarow constructor safer
When I wrote construct_object_item_data_row, I allowed it to truncate
an overlong name by scribbling a string terminator into the name
passed to it and then copying that newly-terminated string into the
data row's callsign field.

This was dangerous if one passed a string literal into the function
for the name, and on at least one platform that led to a segfault.  To
step around that segfault, I made the unit test for the truncation
feature pass a char array instead of a string literal.  That was a
kludge.

This commit fixes construct_object_item_data_row to leave the argument
string it is passed alone.  If it determines that it needs to
truncate, it copies the name into a newly allocated char
array (allocated to be exactly big enough to fit the name and its
terminator), then scribble a new terminator into that, then copy it
into the data row.

It also changes the unit test to pass in a string literal instead of a
char array, to make sure that the new modification works correctly and
doesn't crash.