Add tests of string/long conversion consistency
This commit adds two unit tests of convert_lat/lon_l2s and
convert_lat/lon_s2l.  These tests check that starting from a string,
converting to long, and converting back to string produces the same
string we started with, and that starting from a long, converting to
string, and back to long also gives what we started with.

This is going to be the sort of unit test I create for other
conversion utils, including the new ones I intend to write.