Add a test of converter from screen coords
util.c currently provides only a function that takes screen coords and
converts them to Xastir coordinates, given current screen size and
extents.

This commit attempts to test this function out by setting the NW and
SE corner variables to something that happens to be my current Xastir
setting and computing a center coordinate, the variable actually used
by convert_screen_to_xastir_coordinates, and X/Y scales (in 1/100
second per pixel).

Unfortunately, there are rounding problems that surface when trying to
use my stated NW/SE corner values, and in fact Xastir winds up
recomputing the NW and SE corners from the center coord and scales, so
after computing scales and center from my corners, I recompute the
corner coordinates from the center and scales, which invariable is a
little off from where they started.

The test then checks that (screen_width/2,screen_height/2) actually
maps to the Xastir coordinates for the center lat/lon, that (0,0) maps
onto the NW corner lat/lon, and that (screen_width,screen_height) maps
onto the SE corner lat/lon.

For good measure, I also try to recompute the screen coordinates of
the corners and center using the formula that will ultimately be used
in convert_xastir_to_screen_coordinates, as yet unwritten, and make
sure they are (0,0), (screen_width/2, screen_height/2),
and (screen_width,screen_height) for the NW, center, and SE.

This checks that the existing convert_screen_to_xastir_coordinates
works as it should (it had better, because it's what we've been using
forever).  It also establishes that the formula that is currently used
in many places all over the map code is in fact the inverse of
convert_screen_to_xastir, provided that the NW and SE corners are
recomputed appropriately from the center coord and scales (which
I *think* I see getting done in Xastir, but honestly it's difficult to
follow how scales and corners are actually handled).

What I'm testing here is consistent with what's being done in Xastir
and will serve as a canary in the coal mine once I set up
convert_xastir_to_screen_coordinates and start unit testing *it*.

This is all in one huge test at the moment, but I think it's ok.  I
might split it up into three separate tests as I progress on this
issue, but for now this does the trick.