Add new function convert_xastir_to_screen_coordinates
This adds a new util function, convert_xastir_to_screen_coordinates
that is an exact copy of what is cut and pasted all over Xastir's map
code.

It also adds a test of this function that at the moment is an exact
copy of the one that tests convert_screen_to_xastir_coordinates except
for replacing the cut/pasted version of the old conversion code with a
call to the new converter.

I am not perfectly happy with this as a test suite because the tests
are not really "unit" tests --- they test lots of things all at once.

But they actually check that the functions work, that they are
consistent with each other (i.e. they are inverse functions and
conversion in both directions reproduces the same data as was input).

I am not sure there is anything to be gained by splitting the two
tests up into multiple tests, each testing one corner or just testing
the functions in one directions.  Every step of the test must pass all
assertions, and a single pass/fail indicates whether or not they're
all working correctly.