Replace cut/paste code with function call
This commit replaces 8 cut/paste copies of code that converts Xastir
coordinates to screen coordinates with the unit-tested function
convert_xastir_to_screen_coordinates.

In ONE place in the entire file, the x and y variables were reduced to
16-bit values with the "l16()" macro in the conversion.  I preserved
that.

In all other cases, the x and y values are reduced to 16-bit values
when they are *used*.

Closes #273