Fix error in consistency check
I created two checks of string->long-string and long->string->long for
lat/lon, but only added one of those tests to the "tests[]" array,
then made another error in the util_tests.at file by calling the one
test that *was* in the tests array twice, with different text saying
what was being done.

This commit fixes that and makes sure both tests are actually being
run as intended.