Extract code for shortened filename
All of the map types do a little dance to get a shortened filename to
use in the status line.  If the map file name is too long, it's
truncated on the left and ".." is prepended.

All of these files had cut/pasted code to do exactly the same thing.
Some files even had the code cut and pasted twice, complete with the
full four line comment.

This commit refactors that code into a single function in util.c, and
replaces all instances of that duplicated code with a function call.

There are also two unit tests of the function, to assure it doesn't
truncate a short filename and does truncate a long one correctly.

Closes #275