Pull code for wx alert shape selection into function
One of the things that draw_shapefile_map can be used for is
displaying a single weather alert by highlighting the shape
corresponding to that alert out of the nws shapefiles.

The first thing that draw_shapefile_map does is to try to find that
shape if a non-null alert_entry pointer is given in the arguments.

That code is very simple, but a lot of lines.  This commit pulls all
those lines down into a function "find_wx_alert_shape" that returns
the record number of the appropriate shape.  If the alert_entry
already has its "index" element set.  If that is not set, we search
through the entire DBF file to find the record that has a
key (provided by the dbfawk processing) that matches.  If we find one,
we store its number in the alert_entry.

If the shape is not found or if the function is passed a null
alert_entry pointer, the function returns -1.

This exactly reproduces what the code was doing before, but takes 1
line in draw_shapefile_map instead of 97 lines (including internal
comments), making it easier to see the forest for the trees.

One other change:  There were several static variables and arrays
declared in draw_shapefile_map that had to be static because pointers
to them are stored in the DBFAWK symbol table and they must never be
local variables that could go out of scope and change pointers.

Because find_wx_alert_shape calls dbfawk functions to process the dbf
entries of the file (in order to search for the key), it must also
have access to those variables (key is one of them).  I expect other
refactor work to need the same, so I moved these variables up to be
file-scoped statics.

Part of work on #278