Move creation of two blocks into function
sType is set to a string based on what sort of shapefile we've been
given.  This string is ONLY used in debug statements.

Move this long switch statement down into a function.

Replace the cleanup and return statements for two of those
cases (unsupported or unknown types) into a single conditional.

Moved setting of "viewportRect" into a function for the sole purpose
of making the draw_shapefile_map function have one line instead of
six (including declaration of temporary variables used only for this
purpose.

Not really related to the main purpose of this branch, but definitely
related to #268