Move alert XBM selection to a function
Since commit 88d579 (in Apr 2002), Xastir has used an XBM file to set
a stipple in the "gc_tint" graphics context.  Prior to that, it used
the "gc_tint" graphics context set to a particular fill color, then used
XSetFunction to "or" this color with the underlying map color.  Commit
88d579 did away with tinting.

Since stippling of alerts first tries to find an XBM whose name
matches the tag in the alert_entry.  If it finds it, it uses it as the
stipple pattern, otherwise it uses a generic "alert.xbm" file.

This location of the correct XBM file is a simple operation but takes
many lines, and is amenable to migration to a simple function to
simplify the appearance and readability of draw_shapefile_map.  This
commit adds such a function and replaces the code in
draw_shapefile_map with a call to it.

In addition to that block of code, the readability of the
draw_shapefile_map was further obscured by a couple of really big
comment blocks (much bigger than the code).  One of them explained how
the alert xbm was selected (this comment now adorns the new function).

The other large comment listed every allowable value for the third
argument of XSetFunction, with notes on how they performed prior to
commit 88d579, when the code used to simply tint the shape.  That
comment is was rendered obsolete by commit 88d579, which actually
replaced the function being set with one of those that the comment
explicitly said NOT to use (not to use as a tinting function, that
is).  I have removed that comment block, which had been in place since
Xastir was first imported into CVS in Feb of 2002.

Having moved the selection of xbm file into a separate function, I
also took the time to simplify its construction of the full pathname,
which used to be done with a bunch of strcpy/strcat/intermediate
termination steps as is common all over Xastir, where a simple
xastir_snprintf would do the trick in one line.