Remove an empty "else" block
A very long time ago when Xastir had hard-coded shapefile rendering
for every type of shapefile it supported, it set things like
"lake_flag", "river_flag", "path_flag" and so on, and set the graphics
context color based on those.

Some time long ago, in commit 85c456, even before DBFAWK, when
shapefile rendering code was till in maps.c, a comment was added to a
block of code "Are these actually used anymore by the code?".

In commits 0dfff2 and 62a8be the block of code to which that comment
referred was then put inside an "ifndef WITH_DBFAWK" but the comment
left outside.  When all those ifdefs went away, the comment remained
--- inside an otherwise empty else block.

I'm removing the comment and the else block containing it.