Insert some comments about rtree usage
I implemented the rtree thing in 2004.  It's worked for a long time
and needed no maintenance since it was unleashed as default.  But that
was a long time ago, and I needed to remind myself what some of the
code in map_shp.c is actually doing.  These comments help.