Tidy up handling of point shapefile stuff
The part of the enormous case statement that pertains to points is so
straightforward it doesn't need much refactoring, but was peppered
with commented-out code and comments that didn't add to the
readability, so I'm tidying it up.  This is orthogonal to weather
alert cleanup, but helps me do that cleanup.