Replace a comment
Back when the code for handling polygon shapes with holes was
implemented, a comment was added to weather alert code "We skip the
hole/fill thing for these?".  That was in 2002.

Yes, we skip the hole/fill thing, because NWS shapefiles don't have
polygons with holes, apparently.  Say *that* instead.