Extract code for setting gps track color/label
A very old feature of Xastir is the ability to download GPS tracks by
using GPSMan as a helper application.  GPSMan can create shapefiles
directly, but creates them with a barebones DBF file (with only name
and comment attributes extracted right from the GPS).

Because this feature was aimed at Search and Rescue or other
public-service type events, it was desirable to have a technique to
make multiple GPS tracks distinct from each other, so a hack was added
in that looked to see if the shapefile was in the "GPS" directory of
the Xastir maps directory structure, and if so, look at the file name
for a color.  The track would be labled with that color and use the
base name of the file with the color stripped off as the label.

The way that worked was you'd download the shapefile from GPSMan,
which would stuff it in your user  "gps" directory as
"Untitled_Track_Red.shp", then you'd move it to the system "GPS"
directory and change its name to reflect the color and label you
wanted.

When Alan implemented DBFAWK he suggested that this *should* have been
doable with dbfawk by adding attributes for color and name to the
dbfawk file, but actually since the files are created by GPSMan we
have no such control unless we modify the files, another can of worms.

But as a compromise solution, in commit 6bc21a I set Xastir up to
write out a bare-bones dbfawk file with the same name as the
downloaded gps shapefile.  To use that method, instead of copying the
downloaded shapefile to the GPS directory and editing its name, you
copied the shapefile and its dbfawk file somewhere else and edited the
dbfawk file to get the color and name you wanted, along with any other
special rendering options.

If Xastir finds a DBFAWK file corresponding to the GPS track, it uses
it irrespective of whether it's in the GPS directory or not --- so the
only way you exercise this old method is if you copy the GPS file to
the GPS directory and DON'T copy its automatically created dbfawk.

The code to handle setting the gps label and color this old way is
trivial, but a lot of lines.  Having it right smack in the middle of
the otherwise relatively straightforward SHPT_ARC handling is a
readability problem.

This commit moves that block of code into a function and replaces the
inline version with a function call.

One note:  The original code used two different versions of the
filename to do its work:  "filename" and "filenm".  The former is the
base name of the file and the latter includes the path.  It used
filename to get the gps_label value by stripping off the color, and it
used filenm to get the color.  But since filename has all the
information needed to do both things, I've made the new function only
have one filename string passed into it, and it uses that one for both
purposes.