Pull SHPT_ARC vertex processing into a function
The guts of a loop over shape object vertices that simply extracted
lat/lon coordinates, converted them to Xastir coords, then converted
to screen coords and stored them in an XPoint array has been extracted
into a function.

This loop now consists of "get a list of screen coordinates for this
object" and "draw that list with XDrawLines".

An unnecessary duplicate "if" was removed and its block consolidated
with the block of the immediately preceding if, which had the same
condition.

By simplifying the guts of the loop to this level it is now almost
trivially obvious how the fix of issue #154 will go --- the current
loop over all vertices will need to become two loops, one over parts
the other over vertices within that part.  The code for labeling the
lines will also have to be inside the parts loop, but I haven't gotten
to those lines to tidy them up yet.

Part of #282.