Refactor get_vertex_screen_coords
In my last commit I created a function "get_vertex_screen_coords" that
computed the screen coordinates of a single shape object's vertex and
added them to an array of XPoints that would be passed to XDrawLines.

I then set about to look at the next block of code and discovered
that *it* was doing the very same conversion of one vertex's
coordinates to screen coordinates, but storing them in a couple of
variables for use in compution later.

So I split up the original get_vertex_screen_coords into
"get_vertex_screen_coords_XPoint" and "get_vertex_screen_coords."  The
latter actually does the conversion, the former calls the latter and
stows the result in the XPoint array, taking care of the indexing for
the next one.

More #282.