Pull label rotation computation into function
The label rotation is computed based on the first two vertices in the
line.  Lots-o-lines, not much to see.  Move it down to a function.

Also, the draw_rotated_label_text calls for gps_flag and !gps_flag
were identical except for angle and color.  Consolidate them into a
single call by factoring out the color and angle.

There is very little left to do to the SHPT_ARC case in
draw_shapefile_map.  Just one section dealing with the label hash.