Complete clean up of SHPT_ARC code
This commit moves the code that adds a new label to the label hash to
a function.

This completes my work cleaning up the SHPT_ARC part of the
draw_shapefile_map switch statement, the subject of issue #282, a
subissue of #274.

This is now clean enough that it should be possible to see exactly how
to fix #154.

Closes #282