Rename a loop variable
The code for SHPT_ARC uses the varible "ring" in its loop over
vertices.

Since "ring" is actually something that pertains to multiple parts of
a SHPT_POLYGON structure, and there may also be multiple parts to a
SHPT_ARC structure (c.f. #154), I don't want to confuse "ring" for
"vertex" here.

Add a new "vertex" variable and use it in the loop over vertices in
the SHPT_ARC code.

This lays some groundwork for fixing #154, in which the loop over
parts will be an outer loop and the loop over vertices of that part an
inner loop.