Consolidate two conditionals
The code for drawing the lines of a SHPT_ARC and the code for plotting
its label were in completely separate conditionals, but the
code in the labeling part had conditionals that were redundant with
the code for the line part.

Consolidate these two blocks inside a single if.

The reason for doing this is that the entire block is going to wind up
having a loop over "parts" of a multipart SHPT_ARC when I fix #154,
and this just makes all that work easier.