Fix improper rendering of multi-part SHPT_ARC shapefiles
Shapefiles of SHPT_ARC can have multiple parts for each shape.  Prior
to this commit, Xastir was rendering these multi-part shape as if they
were one big feature.

In issue #152 (trying to find a better default map) it was discovered
that a prime candidate for a better default map was the
"natural-earth-vector" collection has lots of these multi-part shapes,
and sometimes they've made a single multipart shape for features with
the same name but in widely different parts of the world.  Xastir
would try to tie them all together, making weird straight lines
between endpoints of rivers in different continents.

This commit fixes this incorrect rendering.

It was rendered almost trivial by the work done on #282.  The addition
of a couple of extra variables, moving a couple of lines of code
around, and an extra loop around the rendering.

One hiccup was that I had to make sure to reset the color for the
lines repeatedly over every part, because if we label in a different
color it got stuck in that color for all parts other than the first.
Ripping the color and line thickness setting into a little function
and calling it  in the loop fixed that.

I also simplified the handling of the special setting of color, line
width, and line pattern for "gps_flag" being set.

Fixes #154

We still need to fix SHPT_POLY and the rendering of polygon labels
before we can really get to the meat of #152, which will certainly
require some polygon shapefiles (like, for country borders).