Pull forward declarations into a header
In order to be able to use utility functions that I add to the end of
map_shp.c in functions that appear earlier in the file, I've been
creating forward declarations of those functions at the top of the
file.

There are now so many that it's an eyesore.  Move all of those forward
declarations into a header file, and include that.