Start breaking SHPT_POLYGON case into functions
Breaking up the enormous "case SHPT_POLYGON:" section of
draw_shapefile_map into functions is necessary to improve readability
and maintainability.

Here we break out the setting of fill style/stipple and the
preprocessing of a shape to determine whether there are holes to deal
with.

There is an enormous block of comments associated with the latter.
I'm not sure it is valuable to keep that block since it lives on in
git history forever, but for now I moved it out of the switch
statement and into a header for the preprocess_shp_polygon_holes
function.