Refactor lat/lon to screen coords functions
A common operation in draw_shapefile_map is a conversion of
coordinates from lat/lon floats into screen coordinates.

I had previously pulled out several of these into functions, but both
referenced shapefile objects.

This commit creates a "convert_ll_to_screen_coords" function that
takes the lat/lon floats and returns the screen coords.

get_vertex_screen_coords now calls this, passing the lat/lon from the
shapefile object to it.  get_vertex_screen_coords_XPoint calls
get_vertex_screen_coords but saves the resulting values in an XPoint
array and deals with all the bookkeeping of which index it goes into
and which index should be used next, and checking that we don't run
over the array length.

It also fixes an error in code that was never exercised, the case
where the conversion fails.  I was not correctly dereferencing a
pointer in that case.  My compiler didn't warn me that I was assigning
zero to the pointer, because that's not illegal.  It would have had no
effect because the zeroed pointer wouldn't actually get passed back to
the caller anyway.