Replace repeated blocks of conversion code with function calls
Several places in the SHPT_POLYGON code had repeated code exactly the
same as that I have already refactored out of the SHPT_ARC code.
Replace those with the utility functions I have previously created.

The SHPT_POLYGON code had a block of code almost identical to one I
recently replaced in SHPT_ARC, converting a "part" of a shapefile
object into an XPoint array.

There were some oddities in it that made no sense, such as trying to
frob the "high_water_mark_index" value twice, once for an actual index
into the array and a second time that must have been cut/pasted from
elsewhere.  Also, it oddly incremented the loop index and tested it
against the loop end condition instead of just having the increment in
the for() statement, where it is already being checked.  I put the
loop index increment in the for loop.

The loop is already using a separate variable to keep track of the
index into the points array, and get_vertex_screen_coords_XPoint
already handles that properly, as well as correct setting and checking
of the high_water_mark_index.

There are more instances of this operation to replace, but I'm
replacing one at a time and testing the code in between to keep myself
honest.

The loop over vertices and storage into an XPoint array can also be
refactored into a utility function, but that is not done yet.  Right
now, the loop remains contains one line, calling a utility function.