Pull code out of loop that shouldn't have been there
In commit 3d69444 I changed a little if/else statement that was inside
a loop to code that pulled data out of an array that was being created
in that loop.

It should have been outside the loop, because it only makes sense to
do that stuff once, after the loop concludes.

Moving that code now.