Refactor loop that converts shape object to screen coord array
There is a common loop in several places that loops over the number of
vertices is a shape part and converts those vertices into screen
coordinates.

This commit creates a function to do that loop and calls it in one of
the places that loop exists.  There are other places, but I need to
stop for the night and this is a good place to stop.

Tested thoroughly by running Xastir with a variety of polyline
shapefiles, since this one loop I've refactored is in the SHPT_ARC
section of the code.