Convert another loop to a call to a utility function
This commit converts a second instance of a loop over vertices in a
shapefile object part into a call of a function that does the same
thing.

This pattern repeats in several places so reducing the repetitive
pattern is helpful.

In doing this, I spotted a bug here, because it is possible for
shapefiles to have objects with "nParts==0" --- these are supposed to
be treated as if nParts==1, but panPartStart will (or may?) be null.
In such a case, the number of vertices in object->nVertices is the
single list of vertices for the whole shape.  The code in SHPT_POLYGON
does not handle nParts==0 at all, because it always loops over rings
with "for (ring = 0 ; ring < object->nParts; ring++)", which will just
make it skip any that have nParts==0.

In my work on the SHPT_ARC multi-part object bug I accounted for that
possibility, but I have NOT fixed it here yet.  That will be part of
this issue's refactor, but not this commit.