Fix erroneous handling when nParts==0
In the SHPT_POLYGON code, when it comes time to draw the actual
polygons, there's a loop over the number of parts in the object, which
can be 0, 1, or more than one.

Xastir handled the cases 1 and above, but ignored any shape with
nParts==0.

According to the shapelib documentation, nParts can be zero, and if it
is the shape is to be treated as if it only has one part, but the
"panPartStart" pointer in the object can be null, so shouldn't be
dereferenced.

In my SHPT_ARC refactor and the fix of the multi-part arc issue, I
took into account this possibility and made sure to do the right thing
for nParts==0 or nParts==1 without trying to reference panPartStart
(which, in the case of nParts==1 would always be one element whose
value was 0).

This commit applies that same fix to the loop that draws the actual
polygons.

In the code that I have NOT refactored, setting up polygon holes, this
issue is a non-issue, because if a polygon has holes then by
definition it has more than 1 part, and so nParts will never be zero.