Move for loop variable increment, fix high_water_mark_index setting
The code that reads a polygon shape's holes and makes an XPoint array
out of them had an odd structure:

```
for (index=first_vertex_of_part;index<end_of_part;)
{
   get coordinates/clip them, check if on screen
   if on screen
   {
      store in points[i]
      i++
   }
   index++;

   bump high_water_mark_index to index if index is larger
   check if index > end_of_part
}
```

Since the loop is already checking index<end_of_part in its condition,
the check is unnecessary, and the increment of the index belongs in
the for() statement.

the index is NOT used as a storage location in the points array, "i"
is.  So the high water mark shouldn't be set based on index, it should
be based on i.

This commit fixes both things.