Move actual polygon drawing into functions
The structure of the code in SHPT_POLYGON that actually draws the
final polygon is essentially this:

```
if (nPoints>3 && (don't skip this polygon) &&
    EITHER: (don't fill) OR (fill AND we're not a hole polygon))
{
   if (don't fill AND we're a hole polygon)
      draw a dashed boundary around the hole
   else if (not a weather alert)
      draw the polygon using either the clipping region or without,
      depending on whether there are holes
   else (if a weather alert)
      draw a polygon into the pixmap_alerts using gc_tint, presuming
      that there won't ever be any holes in alert regions
   else if (we're filling)
      do the same thing as (! weather alert)
   else
      just draw the boundary
}

This commit makes it look that simple by factoring out the actual
calls to X routines that do the trick.  The code had been obfuscated
slighty by having duplicate blocks in an if (polygon_hole_flag) that
differed only by which graphics context was being used.  Now, the
function I call takes a GC as argument, and we pass in the appropriate
one based on polygon_hole_flag.

HOWEVER:  look at the if/else ifs.   There is one case that catches
"not a weather alert" and the next else catches "is a weather alert",
so anything that doesn't get caught by the first if gets caught in one
of these two elseifs.  Thus, the "(we're filling)" and final else
blocks can never be reached.

I have NOT deleted those superfluous blocks in this commit, but I will
shortly.