Remove unreachable code
Consider this conditional
```
if (a)
 stuff
else if (!b)
 more stuff
else if (b)
 still more stuff
```

This leaves absolutely nothing to be caught by an else --- b and !b
covers everything not covered by the first conditional.

Yet we had another else if and an else here with code in them.
Probably relics of pre-dbfawk stuff.

I've removed the redundant else if, and replaced the contents of an
else with an fprintf.

The else will probably be removed before I'm done with this refactor.