Pull polygon label point selection into function
Because we might need to consider other ways of choosing a polygon
label point (c.f. #157, #154), I have pulled the choice of label point
into a function.  Currently, it just returns the center of the shape's
bounding box, which is what we have always used.

This commit also replaces the "convert lat lon to screen coordinates"
pattern with a call to convert_ll_to_screen_coords, as I've done
elsewhere.